Add files via upload
Alright guys this is the starting point for the CA5 project. Jay took care of the first piece of SQL. And dermot oversaw the beginning of the projects launch up.
The linking to the database in myphpAdmine is complete.
All thats left is to finish sorting the code errors and we are good to go.
There will be errors though.